Moves wait_result.hpp out into include/flow
