+++ library/source/flow/wait_result.cpp
@@ -1,8 +1,1 @@
-//
-//  wait_result.cpp
-//  library
-//
-//  Created by Louis Langholtz on 4/28/23.
-//
-
-#include "wait_result.hpp"
+#include "flow/wait_result.hpp"
